[flutter_tools] remove globals from flutter web platform
