+++ packages/flutter_tools/test/general.shard/web/golden_comparator_process_test.dart
@@ -5,8 +5,8 @@
 import 'dart:convert';
 
 import 'package:flutter_tools/src/base/file_system.dart';
-import 'package:flutter_tools/src/test/flutter_web_platform.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
+import 'package:flutter_tools/src/test/flutter_web_goldens.dart';
 
 import '../../src/common.dart';
 import '../../src/mocks.dart';
